Fixed typos
This is very good feedback so far - and it's great calling out all these things this early in the process! Really appreciate you taking the time to write this. On a positive note, IMO, there is nothing really unexpected or really controversial so far in the feedback that has been given - so I'm optimistic that our views are aligned for this API. Looking forward to reading the last remaining bits of feedback!