Device-Bound Session Credentials Analysis
This is my analysis of DBSC, which suggests a different approach to the
design that is more congruent with the way that cookies operate today
and uses existing technology.

I propose that this be the substance of the TAG's response to #1052,
and that we close that issue (for now) with an "unsatisfied" resolution.