Review feedback
Apologies to those who reviewed in the past.  Adding wrapping is the
right thing to do, but it blows away all the comments.

I've tried to address the feedback about clarity, so there are parts of
this that are rewritten.  In particular, there is a new section on
hiding latency that doesn't claim superior latency in the proposed
alternative.  Instead, it focuses on certainty for servers.

New text talks about the obvious reduction in latency that can be
achieved.  Then it talks about the non-obvious ways in which the
original proposal was able to hide latency and how the alternative can
use the same tricks, but in a more flexible way.