Device-Bound Session Credentials Analysis (#1094)
* Device-Bound Session Credentials Analysis

This is my analysis of DBSC, which suggests a different approach to the
design that is more congruent with the way that cookies operate today
and uses existing technology.

I propose that this be the substance of the TAG's response to #1052,
and that we close that issue (for now) with an "unsatisfied" resolution.

* jyasskin's edits and a little more

Co-authored-by: Jeffrey Yasskin <jyasskin@google.com>

* Review feedback

Apologies to those who reviewed in the past.  Adding wrapping is the
right thing to do, but it blows away all the comments.

I've tried to address the feedback about clarity, so there are parts of
this that are rewritten.  In particular, there is a new section on
hiding latency that doesn't claim superior latency in the proposed
alternative.  Instead, it focuses on certainty for servers.

New text talks about the obvious reduction in latency that can be
achieved.  Then it talks about the non-obvious ways in which the
original proposal was able to hide latency and how the alternative can
use the same tricks, but in a more flexible way.

* Add /sign to /login in the example

* whoops

Co-authored-by: Jeffrey Yasskin <jyasskin@google.com>

* Remove dead text.

* It's redundant, but we don't need to say that

* I should really get better at regexes

* Apply suggestions from code review

Co-authored-by: Jeffrey Yasskin <jyasskin@google.com>

* Rework headings

* nits

Co-authored-by: Jeffrey Yasskin <jyasskin@google.com>

* x on the spot

* Clearing state section

---------

Co-authored-by: Jeffrey Yasskin <jyasskin@google.com>